Holdings no longer show as 4.999999 but 5 - todo with bigdecimal

+++ app/src/main/java/com/jonnycaley/cryptomanager/data/model/DataBase/Holding.kt
@@ -1,5 +1,6 @@
 package com.jonnycaley.cryptomanager.data.model.DataBase
 
 import java.io.Serializable
+import java.math.BigDecimal
 
-data class Holding(var symbol : String, var quantity : Float, var costUsd : Double, var costBtc : Double, var costEth : Double,  var type : String, var imageUrl : String?) : Serializable
+data class Holding(var symbol : String, var quantity : BigDecimal, var costUsd : Double, var costBtc : Double, var costEth : Double,  var type : String, var imageUrl : String?) : Serializable
